Start a thread in the server to subscribe to a rabbitmq queue and process messages
