+++ src/murfey/server/__init__.py
@@ -5,16 +5,27 @@ import functools
 import logging
 import os
 import socket
+from functools import singledispatch
+from threading import Thread
 from typing import Any
 
 import uvicorn
 import workflows
 import zocalo.configuration
 from fastapi.templating import Jinja2Templates
+from ispyb.sqlalchemy._auto_db_schema import (
+    AutoProcProgram,
+    Base,
+    DataCollection,
+    ProcessingJob,
+)
 from rich.logging import RichHandler
+from sqlalchemy.exc import SQLAlchemyError
 
 import murfey
 import murfey.server.ispyb
+from murfey.server.ispyb import DB
+from murfey.util.state import global_state
 
 try:
     from importlib.resources import files  # type: ignore
@@ -99,6 +110,10 @@ def run():
         type=int,
         default=8000,
     )
+    parser.add_argument(
+        "--demo",
+        action="store_true",
+    )
 
     verbosity = parser.add_mutually_exclusive_group()
     verbosity.add_argument(
@@ -125,6 +140,10 @@ def run():
 
     # Set up logging now that the desired verbosity is known
     _set_up_logging(quiet=args.quiet, verbosity=args.verbose)
+
+    rabbit_thread = Thread(target=feedback_listen, daemon=True)
+    logger.info("Starting Murfey RabbitMQ thread")
+    rabbit_thread.start()
 
     logger.info(
         f"Starting Murfey server version {murfey.__version__} for beamline {get_microscope()}, listening on {args.host}:{args.port}"
@@ -225,3 +244,64 @@ def _set_up_logging(quiet: bool, verbosity: int):
 def _set_up_transport(transport_type):
     global _transport_object
     _transport_object = murfey.server.ispyb.TransportManager(transport_type)
+
+
+def feedback_callback(header: dict, message: dict):
+    record = None
+    if message["register"] == "motion_corrected":
+        if global_state.get("motion_corrected") and isinstance(
+            global_state["motion_corrected"], list
+        ):
+            global_state["motion_corrected"].append(message["movie"])
+        else:
+            global_state["motion_corrected"] = [message["movie"]]
+    elif message["register"] == "data_collection":
+        record = DataCollection(imageDirectory=message["image_directory"])
+    elif message["register"] == "processing_job":
+        record = ProcessingJob(
+            dataCollectionId=message["data_collection_id"], recipe=message["recipe"]
+        )
+    elif message["register"] == "auto_proc_program":
+        record = AutoProcProgram(processingJobId=message["processing_job_id"])
+    if record:
+        _register(record, header)
+    elif _transport_object:
+        _transport_object.transport.nack(header, requeue=False)
+
+
+@singledispatch
+def _register(record, header: dict):
+    raise NotImplementedError(f"Not method to register {record} or type {type(record)}")
+
+
+@_register.register
+def _(record: Base, header: dict):
+    if not _transport_object:
+        logger.error(
+            f"No transport object found when processing record {record}. Message header: {header}"
+        )
+        return None
+    try:
+        DB.add(record)
+        DB.commit()
+        _transport_object.transport.ack(header)
+        return getattr(record, record.__table__.primary_key.columns[0].name)
+    except SQLAlchemyError as e:
+        logger.error(f"Murfey failed to insert ISPyB record {record}", e, exc_info=True)
+        _transport_object.transport.nack(header)
+        return None
+    except AttributeError as e:
+        logger.error(
+            f"Murfey could not find primary key when inserting record {record}",
+            e,
+            exc_info=True,
+        )
+        _transport_object.transport.nack(header)
+        return None
+
+
+def feedback_listen():
+    if _transport_object:
+        _transport_object.transport.subscribe(
+            "murfey_feedback", feedback_callback, acknowledgement=True
+        )
